Move certutil code to graylog-server
