+++ data-node/src/main/java/org/graylog/datanode/bootstrap/Main.java
@@ -21,9 +21,9 @@ import com.github.rvesse.airline.builder.CliBuilder;
 import com.google.common.collect.ImmutableSet;
 import org.graylog.datanode.bootstrap.commands.CliCommandHelp;
 import org.graylog.datanode.bootstrap.commands.ShowVersion;
-import org.graylog.datanode.bootstrap.commands.certutil.CertutilCa;
-import org.graylog.datanode.bootstrap.commands.certutil.CertutilCert;
-import org.graylog.datanode.bootstrap.commands.certutil.CertutilHttp;
+import org.graylog.security.certutil.CertutilCa;
+import org.graylog.security.certutil.CertutilCert;
+import org.graylog.security.certutil.CertutilHttp;
 import org.graylog2.bootstrap.CliCommand;
 import org.graylog2.bootstrap.CliCommandsProvider;
