+++ data-node/src/test/java/org/graylog/datanode/integration/DatanodeSecuritySetupIT.java
@@ -24,11 +24,11 @@ import com.github.rholder.retry.WaitStrategies;
 import io.restassured.RestAssured;
 import io.restassured.response.ValidatableResponse;
 import org.apache.http.NoHttpResponseException;
-import org.graylog.datanode.bootstrap.commands.certutil.CertutilCa;
-import org.graylog.datanode.bootstrap.commands.certutil.CertutilCert;
-import org.graylog.datanode.bootstrap.commands.certutil.CertutilHttp;
-import org.graylog.datanode.bootstrap.commands.certutil.TestableConsole;
 import org.graylog.datanode.testinfra.DatanodeContainerizedBackend;
+import org.graylog.security.certutil.CertutilCa;
+import org.graylog.security.certutil.CertutilCert;
+import org.graylog.security.certutil.CertutilHttp;
+import org.graylog.security.certutil.console.TestableConsole;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
