+++ graylog2-server/src/main/java/org/graylog/security/certutil/console/CommandLineConsole.java
@@ -14,7 +14,7 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog.datanode.bootstrap.commands.certutil;
+package org.graylog.security.certutil.console;
 
 import java.util.Locale;
