+++ graylog2-server/src/test/java/org/graylog2/security/certutil/CertutilCaTest.java
@@ -14,9 +14,11 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog.datanode.bootstrap.commands.certutil;
+package org.graylog2.security.certutil;
 
 import org.assertj.core.api.Assertions;
+import org.graylog.security.certutil.CertutilCa;
+import org.graylog.security.certutil.console.TestableConsole;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
