+++ graylog2-server/src/test/java/org/graylog2/security/certutil/CertutilHttpTest.java
@@ -14,10 +14,13 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog.datanode.bootstrap.commands.certutil;
+package org.graylog2.security.certutil;
 
 import org.assertj.core.api.Assertions;
-import org.bouncycastle.asn1.x509.Extension;
+import org.graylog.security.certutil.CertutilCa;
+import org.graylog.security.certutil.CertutilCert;
+import org.graylog.security.certutil.CertutilHttp;
+import org.graylog.security.certutil.console.TestableConsole;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
